cast to UINT8 first to prevent sign extension
date	2001.12.03.21.21.00;	author rmoore1;	state Exp;

+++ source/components/namespace/nsutils.c
@@ -2,7 +2,7 @@
  *
  * Module Name: nsutils - Utilities for accessing ACPI namespace, accessing
  *                        parents and siblings and Scope manipulation
- *              $Revision: 1.95 $
+ *              $Revision: 1.96 $
  *
  *****************************************************************************/
 
@@ -596,7 +596,7 @@ AcpiNsExternalizeName (
             /* <count> 4-byte names */
 
             NamesIndex = PrefixLength + 2;
-            NumSegments = (UINT32) InternalName[PrefixLength + 1];
+            NumSegments = (UINT32) (UINT8) InternalName[PrefixLength + 1];
             break;
